Added interfaces and services, removed comments.
